Refactored the code and documented it
