+++ cmd/main.go
@@ -6,6 +6,7 @@ import (
 
 	"log"
 
+	"github.com/sergio-prgm/tf-module/pkg/gen"
 	"github.com/sergio-prgm/tf-module/pkg/inout"
 	"github.com/sergio-prgm/tf-module/pkg/scf"
 	"github.com/sergio-prgm/tf-module/pkg/util"
@@ -30,13 +31,16 @@ func main() {
 
 	parsedBlocks := inout.ReadTfFiles(src)
 	resourcesMapping := inout.JsonParser(src + "aztfexportResourceMapping.json")
-	csv_resources := inout.ParseCSV(yml + "module_map.csv")
+	/*
+		csv_resources := inout.ParseCSV(yml + "module_map.csv")
 
-	mapped_yaml := inout.GenerateModuleYaml(resourcesMapping, csv_resources)
-	inout.WriteYaml(yml+"tfmodule.yaml", mapped_yaml)
+		mapped_yaml := gen.GenerateModuleYaml(resourcesMapping, csv_resources)
+
+		inout.WriteYaml(yml+"tfmodule.yaml", mapped_yaml)
+	*/
 	configModules := inout.ReadConfig(yml)
 
-	resourceMap := inout.CreateVars(parsedBlocks.Resources, configModules.Modules)
+	resourceMap := gen.CreateVars(parsedBlocks.Resources, configModules.Modules)
 
 	scf.CreateFolders(configModules)
 	err := scf.CreateFiles(parsedBlocks, resourceMap, configModules)
@@ -44,6 +48,6 @@ func main() {
 		log.Fatal(err)
 	}
 
-	inout.GenerateImports(resourcesMapping, configModules)
+	gen.GenerateImports(resourcesMapping, configModules)
 
 }
